rildixon - rocm_ais: cast devId to int in aisXferReq init
nixlBasicDesc::devId is uint64_t while aisXferReq::dev_id is int (hipSetDevice
takes int), so braced aggregate init is a narrowing conversion that
-Werror=narrowing rejects. Caught by compiling in the nixl-rocm image.

+++ src/plugins/rocm_ais/ais_backend.cpp
@@ -199,7 +199,7 @@ nixlAisEngine::prepXfer(const nixl_xfer_op_t &operation,
                                   (size_t)file_desc.addr,
                                   file_handle->hip_fhandle,
                                   (operation == NIXL_READ) ? hipFileBatchRead : hipFileBatchWrite,
-                                  mem_desc.devId});
+                                  static_cast<int>(mem_desc.devId)});
     }
 
     if (reqs.empty()) {
